updated bugs with JS and failing to load plugins
