+++ app/javascript/packs/application.js
@@ -4,6 +4,9 @@
 // that code so it'll be compiled.
 
 import Rails from "@rails/ujs"
+require("turbolinks").start()
+Rails.start()
+ActiveStorage.start()
 // import Turbolinks from "turbolinks"
 import * as ActiveStorage from "@rails/activestorage"
 import "channels"
@@ -21,26 +24,11 @@ import {background_page_height} from "../plugins/init_background_height";
 document.addEventListener('turbolinks:load', () => {
   // Call your JS functions here
   initAutocomplete();
-});
-
-document.addEventListener('turbolinks:load', () => {
-  // Call your functions here, e.g:
   initMapbox();
-});
 
-$(function() {
   background_page_height()
-  const banner_exists = document.querySelector('#banner')
-  const header_logo = document.querySelector('#logo');
-  const header_links = document.querySelector('#links')
-  if (banner_exists) {
-    header_logo.classList.add('hidden')
-    header_links.classList.add('hidden')
-    LogoTogal();
-  }
 
 
-  Rails.start()
-  // Turbolinks.start()
-  ActiveStorage.start()
-})
+
+  LogoTogal();
+});
